Create the Game Class
This class will be the main class used to play Minesweeper.
The constructor accepts  parameters: numberOfRows, numberOfColumns, numberOfBombs.
Inside have property called _board which will be an instance of Board class.
And a method .playMove() that will include all of the functionality needed
to play a session of Minesweeper, including flipping a tile,
letting the user know if they discovered a bomb,
and allowing a user to continue otherwise (until they win, or lose).